Imp: MinGW32 コンパイル対応 upatchid:271 Contributed by　minoki
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk2@2543 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/convert/convert_util2.h
@@ -34,7 +34,6 @@
 #include "mem/CMemory.h"
 
 
-
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 //
 //    Quoted-Printable デコード
@@ -526,7 +525,7 @@ bool CheckUUHeader( const CHAR_TYPE *pSrc, const int nLen, TCHAR *pszFilename )
 	}
 	// ファイル名を格納
 	if( pszFilename ){
-		strtotcs( pszFilename, pwstart, nwlen );
+		strtotcs( pszFilename, pwstart, (size_t)nwlen );
 		pszFilename[nwlen] = _WINT('\0');
 	}
